+++ siopv2/src/authorization_request.rs
@@ -1,4 +1,4 @@
-use crate::{openid4vc_extension::SIOPv2, ClaimRequests, StandardClaimsRequests};
+use crate::{siopv2::SIOPv2, ClaimRequests, StandardClaimsRequests};
 use anyhow::{anyhow, Result};
 use is_empty::IsEmpty;
 use oid4vc_core::builder_fn;
